# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Tajweed Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /announcements:
    get:
      operationId: listAnnouncements
      tags: [announcements]
      summary: List announcements
      parameters:
        - name: courseId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Announcements list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Announcement"
    post:
      operationId: createAnnouncement
      tags: [announcements]
      summary: Create announcement (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"

  # ── Certificates ──────────────────────────────────────────────────────────────
  /certificates:
    get:
      operationId: listCertificates
      tags: [certificates]
      summary: List certificates (admin sees all, student sees own)
      responses:
        "200":
          description: Certificates list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Certificate"
    post:
      operationId: createCertificate
      tags: [certificates]
      summary: Issue certificate (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CertificateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Certificate"

  # ── Payments ──────────────────────────────────────────────────────────────────
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List payments (admin sees all, student sees own)
      responses:
        "200":
          description: Payments list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Record payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}:
    patch:
      operationId: updatePayment
      tags: [payments]
      summary: Update payment status (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"

  # ── Applications ──────────────────────────────────────────────────────────────
  /applications:
    get:
      operationId: listApplications
      tags: [applications]
      summary: List applications (admin only)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
      responses:
        "200":
          description: Applications list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
    post:
      operationId: createApplication
      tags: [applications]
      summary: Submit application (public, no auth required)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "201":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{id}:
    patch:
      operationId: updateApplication
      tags: [applications]
      summary: Approve or reject application (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"

  # ── Progress ──────────────────────────────────────────────────────────────────
  /progress:
    get:
      operationId: getStudentProgress
      tags: [progress]
      summary: Get student progress (mastery per rule)
      parameters:
        - name: studentId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Progress list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProgressEntry"
  /progress/{ruleId}:
    patch:
      operationId: updateProgress
      tags: [progress]
      summary: Update mastery for a rule
      parameters:
        - name: ruleId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgressUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgressEntry"

  # ── Daily Wird ────────────────────────────────────────────────────────────────
  /daily-wird:
    get:
      operationId: getDailyWird
      tags: [dailyWird]
      summary: Get today's daily wird content
      responses:
        "200":
          description: Daily wird
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyWird"
    post:
      operationId: createDailyWird
      tags: [dailyWird]
      summary: Create daily wird entry (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DailyWirdInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyWird"

  # ── Weekly Star ────────────────────────────────────────────────────────────────
  /weekly-star:
    get:
      operationId: getWeeklyStar
      tags: [weeklyStar]
      summary: Get current weekly star student
      responses:
        "200":
          description: Weekly star
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeeklyStar"
    post:
      operationId: setWeeklyStar
      tags: [weeklyStar]
      summary: Set weekly star (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WeeklyStarInput"
      responses:
        "201":
          description: Set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeeklyStar"

  # ── Comment Templates ──────────────────────────────────────────────────────────
  /comment-templates:
    get:
      operationId: listCommentTemplates
      tags: [commentTemplates]
      summary: List comment templates (admin only)
      responses:
        "200":
          description: Templates list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CommentTemplate"
    post:
      operationId: createCommentTemplate
      tags: [commentTemplates]
      summary: Create comment template (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentTemplateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentTemplate"
  /comment-templates/{id}:
    delete:
      operationId: deleteCommentTemplate
      tags: [commentTemplates]
      summary: Delete comment template (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Audio Library ──────────────────────────────────────────────────────────────
  /audio-library:
    get:
      operationId: listAudioLibrary
      tags: [audioLibrary]
      summary: List audio library entries (professional reciters)
      parameters:
        - name: ruleId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Audio library list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AudioLibraryItem"
    post:
      operationId: createAudioLibraryItem
      tags: [audioLibrary]
      summary: Add audio library item (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AudioLibraryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AudioLibraryItem"
  /audio-library/{id}:
    delete:
      operationId: deleteAudioLibraryItem
      tags: [audioLibrary]
      summary: Delete audio library item (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Student Notes ──────────────────────────────────────────────────────────────
  /students/{studentId}/notes:
    get:
      operationId: listStudentNotes
      tags: [studentNotes]
      summary: List private notes for a student (admin only)
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notes list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudentNote"
    post:
      operationId: createStudentNote
      tags: [studentNotes]
      summary: Add private note for a student (admin only)
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentNoteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentNote"

  # ── Dashboard ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Announcement:
      type: object
      properties:
        id:
          type: integer
        courseId:
          type: ["integer", "null"]
        title:
          type: string
        content:
          type: string
        createdAt:
          type: string
      required: [id, title, content, createdAt]
    AnnouncementInput:
      type: object
      properties:
        courseId:
          type: ["integer", "null"]
        title:
          type: string
        content:
          type: string
      required: [title, content]

    # ── Certificates ──────────────────────────────────────────────────────────
    Certificate:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        courseId:
          type: integer
        issuedAt:
          type: string
        certificateUrl:
          type: ["string", "null"]
        student:
          $ref: "#/components/schemas/User"
        course:
          $ref: "#/components/schemas/Course"
      required: [id, studentId, courseId, issuedAt]
    CertificateInput:
      type: object
      properties:
        studentId:
          type: integer
        courseId:
          type: integer
        certificateUrl:
          type: ["string", "null"]
      required: [studentId, courseId]

    # ── Payments ──────────────────────────────────────────────────────────────
    Payment:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        courseId:
          type: integer
        amount:
          type: number
        status:
          type: string
          enum: [pending, paid, cancelled, refunded]
        method:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        paidAt:
          type: ["string", "null"]
        student:
          $ref: "#/components/schemas/User"
        course:
          $ref: "#/components/schemas/Course"
        createdAt:
          type: string
      required: [id, studentId, courseId, amount, status, createdAt]
    PaymentInput:
      type: object
      properties:
        studentId:
          type: integer
        courseId:
          type: integer
        amount:
          type: number
        method:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required: [studentId, courseId, amount]
    PaymentUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, paid, cancelled, refunded]
        method:
          type: string
        notes:
          type: string
        paidAt:
          type: ["string", "null"]

    # ── Applications ──────────────────────────────────────────────────────────
    Application:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        desiredLevel:
          type: integer
        message:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected]
        createdAt:
          type: string
      required: [id, name, email, desiredLevel, status, createdAt]
    ApplicationInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        desiredLevel:
          type: integer
        message:
          type: ["string", "null"]
      required: [name, email, desiredLevel]
    ApplicationUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, approved, rejected]
      required: [status]

    # ── Progress ──────────────────────────────────────────────────────────────
    ProgressEntry:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        ruleId:
          type: integer
        masteryLevel:
          type: integer
          description: "0=not started, 1=in progress, 2=mastered"
        selfAssessment:
          type: integer
          description: "1-5 student self-rating"
        updatedAt:
          type: string
        rule:
          $ref: "#/components/schemas/TajweedRule"
      required: [id, studentId, ruleId, masteryLevel, updatedAt]
    ProgressUpdate:
      type: object
      properties:
        masteryLevel:
          type: integer
        selfAssessment:
          type: integer

    # ── Daily Wird ─────────────────────────────────────────────────────────────
    DailyWird:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
        ruleId:
          type: ["integer", "null"]
        title:
          type: string
        content:
          type: string
        rule:
          $ref: "#/components/schemas/TajweedRule"
      required: [id, date, title, content]
    DailyWirdInput:
      type: object
      properties:
        date:
          type: string
        ruleId:
          type: ["integer", "null"]
        title:
          type: string
        content:
          type: string
      required: [date, title, content]

    # ── Weekly Star ────────────────────────────────────────────────────────────
...
    StudentNoteInput:
      type: object
      properties:
        note:
          type: string
      required: [note]

    # ── Dashboard ─────────────────────────────────────────────────────────────
...